refact & rename : 봉사 카드 데이터 클래스를 Volunteer로 변경 / Volunteer 클래스를 share 패키지로 이동
