--- app/src/main/java/com/example/potatoservice/ui/home/Service.kt
@@ -1,12 +1,0 @@
-package com.example.potatoservice.ui.home
-
-data class Service(
-	val title:String,
-	val organization:String,
-	val category:String,
-	val recruitmentPeriod:String,
-	val recruitmentNumber:String,
-	val time:String,
-	val serviceRecognitionTime:String,
-	val location:String
-)
