+++ app/src/main/java/com/example/potatoservice/ui/mypage/MyPageFragment.kt
@@ -8,6 +8,7 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.potatoservice.databinding.FragmentMypageBinding
+import com.example.potatoservice.ui.share.Volunteer
 
 class MyPageFragment : Fragment() {
