+++ app/src/main/java/com/example/potatoservice/ui/share/AdapterCallback.kt
@@ -1,7 +1,5 @@
 package com.example.potatoservice.ui.share
 
-import com.example.potatoservice.ui.home.Service
-
 interface AdapterCallback {
-	fun onClicked(service: Service)
+	fun onClicked(volunteer: Volunteer)
 }
